+++ src/main/java/br/com/schumaker/octopus/framework/ioc/Environment.java
@@ -4,9 +4,9 @@ import br.com.schumaker.octopus.framework.io.PropertiesReader;
 
 import java.util.Properties;
 
-import static br.com.schumaker.octopus.framework.reflection.AppProperties.DEFAULT_VALUE_NAME;
-import static br.com.schumaker.octopus.framework.reflection.AppProperties.SERVER_CONTEXT;
-import static br.com.schumaker.octopus.framework.reflection.AppProperties.SERVER_PORT;
+import static br.com.schumaker.octopus.framework.ioc.Environment.AppProperties.DEFAULT_VALUE_NAME;
+import static br.com.schumaker.octopus.framework.ioc.Environment.AppProperties.SERVER_CONTEXT;
+import static br.com.schumaker.octopus.framework.ioc.Environment.AppProperties.SERVER_PORT;
 
 public class Environment {
     private static final String DEFAULT_VALUE_VALUE = "0";
@@ -34,5 +34,19 @@ public class Environment {
 
     public String getServerContext() {
         return properties.getProperty(SERVER_CONTEXT, SERVER_CONTEXT_DEFAULT);
+    }
+
+    public static class AppProperties {
+
+        public static final String APP_NAME = "Octopus";
+        public static final String APP_VERSION = "0.0.0";
+        public static final String SERVER_PORT = "server.port";
+        public static final String SERVER_CONTEXT = "server.context";
+        public static final String DB_DRIVER = "db.driver";
+        public static final String DB_URL = "db.url";
+        public static final String DB_USER = "db.user";
+        public static final String DB_PASSWORD = "db.password";
+
+        public static final String DEFAULT_VALUE_NAME = "oc.annotation.default.value";
     }
 }
