+++ src/main/java/br/com/schumaker/octopus/framework/jdbc/DbConnection.java
@@ -1,7 +1,6 @@
 package br.com.schumaker.octopus.framework.jdbc;
 
 import br.com.schumaker.octopus.framework.ioc.Environment;
-import br.com.schumaker.octopus.framework.reflection.AppProperties;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -13,9 +12,9 @@ public class DbConnection {
         try {
             Class.forName("com.mysql.cj.jdbc.Driver");
             return DriverManager.getConnection(
-                    environment.getKey(AppProperties.DB_URL),
-                    environment.getKey(AppProperties.DB_USER),
-                    environment.getKey(AppProperties.DB_PASSWORD));
+                    environment.getKey(Environment.AppProperties.DB_URL),
+                    environment.getKey(Environment.AppProperties.DB_USER),
+                    environment.getKey(Environment.AppProperties.DB_PASSWORD));
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
