--- src/main/java/br/com/schumaker/octopus/framework/reflection/AppProperties.java
@@ -1,15 +1,0 @@
-package br.com.schumaker.octopus.framework.reflection;
-
-public class AppProperties {
-
-    public static final String APP_NAME = "Octopus";
-    public static final String APP_VERSION = "0.0.0";
-    public static final String SERVER_PORT = "server.port";
-    public static final String SERVER_CONTEXT = "server.context";
-    public static final String DB_DRIVER = "db.driver";
-    public static final String DB_URL = "db.url";
-    public static final String DB_USER = "db.user";
-    public static final String DB_PASSWORD = "db.password";
-
-    public static final String DEFAULT_VALUE_NAME = "oc.annotation.default.value";
-}
