+++ src/main/resources/application-local.properties
@@ -1,0 +1,8 @@
+server.port=80
+server.context=/product-api
+db.url=jdbc:mysql://db:3306/octopus
+db.user=root
+db.password=tester
+product.name=beer
+product.price=2
+product.qtd=true
